rename utils to helpers according to generator template best practises
