+++ template/src/api/index.js
@@ -1,5 +1,5 @@
 import { File } from "@asyncapi/generator-react-sdk";
-import { capitalize, getProtocol, getConfig, camelCase, convertToFilename, convertOpertionIdToMiddlewareFn } from "../../../utils/index";
+import { capitalize, getProtocol, getConfig, camelCase, convertToFilename, convertOpertionIdToMiddlewareFn } from "../../../helpers/index";
 
 export default function indexEntrypointFile({asyncapi, params}) {
     const protocol = asyncapi.server(params.server).protocol() === 'mqtts' ? 'mqtt' : asyncapi.server(params.server).protocol();
